pyproject.toml
[project]
name = "devmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
]
